requirements.txt
matplotlib==3.10.0
nltk==3.9.1
numpy==2.0.2
PyYAML==6.0.2
torch==2.5.1
tqdm==4.67.1
umap-learn==0.5.7

